requirements.txt
torch==1.8.1
numpy==1.19.5
pandas==1.1.4
tqdm==4.56.2
PyYAML==5.4.1
scikit_learn==1.0
